made a spellbook UI
https://www.youtube.com/watch?v=z5_UdEcFMOc
used this youtube video to make the base for the spellbook, and then I added an array to track what pages the player collected and labelled accordingly by looking at jason's implementation of the dialogue script. Still trying to send a signal between the book_ui.gd script and the pickup script so that the page can actually be added to the spellbook. player can collect pages now