added death scene and dying animation
I finally succeeded in using the signal thing in godot because i found out that the signal between two scripts has to be in the same scene from the beginning, so i dropped the death signal from the time system node to the forest scene and somehow it worked.
and it made it easier to add in the death scene now that the forest node is controlling the time.
also found out that the way to change colors is labeled under default color instead of just color >:(